lots of internal tag handler changes.
bars, roundtime, and compass now all use the IProperty interface, so you will need
to register yourself as an IPropertyListener to get changes. 

there is also now a document tag handler that handles the chars of the main "document"
created by the stormfrontstream. this is a hack for now to get more fine grained control
of how characters are appended to the client.. will probably have to revisit this eventually

git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@67 bb130017-240f-0410-9613-9c2c827820fb